+++ plugin.video.haystacktv/resources/language/resource.language.he_il/strings.po
@@ -7,22 +7,22 @@ msgstr ""
 "Project-Id-Version: plugin.video.haystacktv\n"
 "Report-Msgid-Bugs-To: http://forum.kodi.tv/showthread.php?tid=316485\n"
 "POT-Creation-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"Last-Translator: Lunatixz Translation Team\n"
-"Language-Team: English (http://www.transifex.com/projects/p/xbmc-addons/language/en/)\n"
+"PO-Revision-Date: 2017-10-30 07:58+0200\n"
+"Last-Translator: A. Dambledore\n"
+"Language-Team: Eng2Heb\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: 8bit\n"
-"Language: en\n"
+"Language: he_IL\n"
 "Plural-Forms: nplurals=2; plural=(n != 1);\n"
 
 msgctxt "#30000"
 msgid "Enable Debugging [Log errors]"
-msgstr ""
+msgstr "אפשר מצב איתור באגים"
 
 msgctxt "#30001"
 msgid "Something went wrong, Try again..."
-msgstr ""
+msgstr "האתר לא זמין. נסה שנית מאוחר יותר..."
 
 msgctxt "#30002"
 msgid "Select Preferred Quality"
