+++ plugin.video.haystacktv/resources/settings.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8" standalone="yes"?>
 <settings>
-    <setting id="Enable_Debugging" type="bool" label="30000" default="false" />
+    <setting id="Enable_Debugging" type="bool" label="30000"  default="false" />
+    <setting id="Quality" type="enum" label="30002"  default="3" values="SD|720p|1080p|Highest Available"/>
 </settings>
